+++ BaseUri.php
@@ -32,6 +32,7 @@ use function end;
 use function explode;
 use function implode;
 use function in_array;
+use function is_bool;
 use function preg_match;
 use function rawurldecode;
 use function str_repeat;
@@ -359,6 +360,26 @@ class BaseUri implements Stringable, JsonSerializable, UriAccess
             },
             $this->uriFactory
         );
+    }
+
+    /**
+     * Apply the callback if the given "condition" is (or resolves to) true.
+     *
+     * @param (callable($this): bool)|bool $condition
+     * @param callable($this): (self|null) $onSuccess
+     * @param ?callable($this): (self|null) $onFail
+     */
+    final public function when(callable|bool $condition, callable $onSuccess, ?callable $onFail = null): self
+    {
+        if (!is_bool($condition)) {
+            $condition = $condition($this);
+        }
+
+        return match (true) {
+            $condition => $onSuccess($this),
+            null !== $onFail => $onFail($this),
+            default => $this,
+        } ?? $this;
     }
 
     final protected function computeOrigin(Psr7UriInterface|UriInterface $uri, ?string $nullValue): Psr7UriInterface|UriInterface|null